Refactor: landingPage footer, container 분리
